Manage field and model
